algorithm/leetcode/double_pointer_lt16_test.go
/**
 * @Author: pei, xingxin
 * @mail: xingxinpei@gmail.com
 * @Description:
 * @File: double_pointer_lt16_test.go
 * @Version: 1.0.0
 * @Data: 2021/7/24 3:34 PM
 */
package leetcode

import "testing"

func Test_threeSumClosest(t *testing.T) {
	type args struct {
		nums   []int
		target int
	}
	tests := []struct {
		name string
		args args
		want int
	}{
		// TODO: Add test cases.
		{
			"simple test",
			args{
				[]int{-1,2,1,-4},
				1,
			},
			2,
		},
		{
			"bad case",
			args{
				[]int{1,1,-1,-1,3},
				-1,
			},
			-1,
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := threeSumClosest(tt.args.nums, tt.args.target); got != tt.want {
				t.Errorf("threeSumClosest() = %v, want %v", got, tt.want)
			}
		})
	}
}
